Simplify BlockFilters file
This removes a file and some unnecessary code
